% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrs.R
\name{GNRS}
\alias{GNRS}
\title{Standardize political division names}
\usage{
GNRS(political_division_dataframe)
}
\arguments{
\item{political_division_dataframe}{A properly formatted dataframe, see http://bien.nceas.ucsb.edu/bien/tools/gnrs/gnrs-api/}
}
\value{
Dataframe containing GNRS results.
}
\description{
GNRS returns standardized political division names (according to geonames.org).
}
\examples{
{
gnrs_testfile <- 
read.csv("http://bien.nceas.ucsb.edu/bien/wp-content/uploads/2019/02/gnrs_testfile.csv")

results <- GNRS(political_division_dataframe = gnrs_testfile)
  

}
}
