% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnrs_template.R
\name{GNRS_template}
\alias{GNRS_template}
\title{Make a template for a GNRS query}
\usage{
GNRS_template(nrow = 1)
}
\arguments{
\item{nrow}{The number of rows to include in the template}
}
\value{
Template data.frame that can be populated and then used in GNRS queries.
}
\description{
GNRS_template builds a template that can be populated to submit a GNRS query.
}
\examples{
\dontrun{

template<-GNRS_template(nrow = 2)
template$country<-c("United Stapes","Mexico")
template$state_province<-c("Arizona","Sinalo")
GNRS(political_division_dataframe = template)  

}
}
